Main R code snippets - can be run together
These are code blocks that can generate the objects used in the analysis of this manuscript. Please combine them together and focus on where the objects are located (e.g., QIIME2 objects) and where the working folder is stored. If there is any problems please contact us at Benjamin.wu@nyulangone.org.